crc dnp

+++ src/include/crc/code_crckrmit.hh
@@ -12,8 +12,8 @@ private:
                              uint8_t pre) const override;
 
   virtual uint64_t result(uint64_t value) const noexcept override {
-    uint16_t lowByte = (value & 0xFF00) >> 8;
-    uint16_t highByte = (value & 0x00FF) << 8;
+    uint16_t lowByte = (static_cast<uint16_t>(value) & 0xFF00) >> 8;
+    uint16_t highByte = (static_cast<uint16_t>(value) & 0x00FF) << 8;
     return lowByte | highByte;
   }
 };
